refactor: split searchUser for clear and presentation component
